requirements.txt
flax==0.6.3
hydra-core==1.3.2
jax==0.3.25
matplotlib==3.6.2
numpy==1.23.5
omegaconf==2.3.0
optax==0.1.4
pandas==1.5.2
seaborn==0.13.0
